Improve dirvish mode-line some more
The code's a bit messy (feels like 1/3rd of the config), and I can't
find a good way to get the number of lines in a target buffer (good
meaning without `save-excursion`), but it's progress.

This changes specifically adjust the line counts to be correct in
dirvish buffers, including parent layouts, omitting the information
that I couldn't sort out above in a clear manner.